+++ lib/src/themes/expression/text_expression.dart
@@ -41,10 +41,33 @@ class LayoutPlacementExpression extends Expression {
   Set<String> properties() => _delegate.properties();
 }
 
+class LayoutJustifyExpression extends Expression {
+  final Expression _delegate;
+
+  LayoutJustifyExpression(this._delegate)
+      : super('justify(${_delegate.cacheKey})');
+
+  LayoutJustify evaluate(EvaluationContext context) {
+    final result = _delegate.evaluate(context);
+    if (result is String) {
+      return LayoutJustify.fromName(result);
+    } else if (result != null) {
+      context.logger.warn(() => 'expected string but got $result');
+    }
+    return LayoutJustify.DEFAULT;
+  }
+
+  @override
+  Set<String> properties() => _delegate.properties();
+}
+
 extension TextExpressionExtension on Expression {
   LayoutAnchorExpression asLayoutAnchorExpression() =>
       LayoutAnchorExpression(this);
 
   LayoutPlacementExpression asLayoutPlacementExpression() =>
       LayoutPlacementExpression(this);
+
+  LayoutJustifyExpression asLayoutJustifyExpression() =>
+      LayoutJustifyExpression(this);
 }