+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.4.0
 
 * text halo color can now use expressions
+* add support for text-justify and text-max-width
 * bug fix to theme filters that reference zoom level
 * breaking API change, `TilesetPreprocessor` now requires a zoom level
 ## 2.3.3